+++ ov/unified.go
@@ -2641,5 +2641,6 @@ func populateLayerFromYAML(layer *Layer, ly *LayerYAML) {
 	layer.vars = ly.Vars
 	layer.tasks = ly.Task
 	layer.apk = ly.Apk
+	layer.reboot = ly.Reboot
 	layer.shell = ly.Shell
 }